refactor(auth): extract login logic to feature module and fix tests
